+++ src/chain.h
@@ -23,11 +23,11 @@ extern CSharedCriticalSection cs_mapBlockIndex;
 class CBlockFileInfo
 {
 public:
-    unsigned int nBlocks; //!< number of blocks stored in file
-    unsigned int nSize; //!< number of used bytes of block file
-    unsigned int nUndoSize; //!< number of used bytes in the undo file
-    unsigned int nHeightFirst; //!< lowest height of block in file
-    unsigned int nHeightLast; //!< highest height of block in file
+    uint32_t nBlocks; //!< number of blocks stored in file
+    uint64_t nSize; //!< number of used bytes of block file
+    uint64_t nUndoSize; //!< number of used bytes in the undo file
+    uint32_t nHeightFirst; //!< lowest height of block in file
+    uint32_t nHeightLast; //!< highest height of block in file
     uint64_t nTimeFirst; //!< earliest time of block in file
     uint64_t nTimeLast; //!< latest time of block in file
